+++ samples/cli/accelbyte_py_sdk_cli/eventlog/_get_event_by_namespace_handler.py
@@ -34,18 +34,18 @@ from accelbyte_py_sdk.api.eventlog.models import ModelsEventResponse
 
 @click.command()
 @click.argument("end_date", type=str)
-@click.argument("page_size", type=float)
+@click.argument("page_size", type=int)
 @click.argument("start_date", type=str)
-@click.option("--offset", "offset", type=float)
+@click.option("--offset", "offset", type=int)
 @click.option("--namespace", type=str)
 @click.option("--login_as", type=click.Choice(["client", "user"], case_sensitive=False))
 @click.option("--login_with_auth", type=str)
 @click.option("--doc", type=bool)
 def get_event_by_namespace_handler(
         end_date: str,
-        page_size: float,
+        page_size: int,
         start_date: str,
-        offset: Optional[float] = None,
+        offset: Optional[int] = None,
         namespace: Optional[str] = None,
         login_as: Optional[str] = None,
         login_with_auth: Optional[str] = None,
