+++ samples/cli/accelbyte_py_sdk_cli/eventlog/_get_user_activities_handler.py
@@ -34,16 +34,16 @@ from accelbyte_py_sdk.api.eventlog.models import ModelsEventResponse
 
 @click.command()
 @click.argument("user_id", type=str)
-@click.argument("page_size", type=float)
-@click.option("--offset", "offset", type=float)
+@click.argument("page_size", type=int)
+@click.option("--offset", "offset", type=int)
 @click.option("--namespace", type=str)
 @click.option("--login_as", type=click.Choice(["client", "user"], case_sensitive=False))
 @click.option("--login_with_auth", type=str)
 @click.option("--doc", type=bool)
 def get_user_activities_handler(
         user_id: str,
-        page_size: float,
-        offset: Optional[float] = None,
+        page_size: int,
+        offset: Optional[int] = None,
         namespace: Optional[str] = None,
         login_as: Optional[str] = None,
         login_with_auth: Optional[str] = None,
