+++ samples/cli/accelbyte_py_sdk_cli/eventlog/_query_event_stream_handler.py
@@ -36,8 +36,8 @@ from accelbyte_py_sdk.api.eventlog.models import ModelsGenericQueryPayload
 @click.command()
 @click.argument("body", type=str)
 @click.option("--end_date", "end_date", type=str)
-@click.option("--offset", "offset", type=float)
-@click.option("--page_size", "page_size", type=float)
+@click.option("--offset", "offset", type=int)
+@click.option("--page_size", "page_size", type=int)
 @click.option("--start_date", "start_date", type=str)
 @click.option("--namespace", type=str)
 @click.option("--login_as", type=click.Choice(["client", "user"], case_sensitive=False))
@@ -46,8 +46,8 @@ from accelbyte_py_sdk.api.eventlog.models import ModelsGenericQueryPayload
 def query_event_stream_handler(
         body: str,
         end_date: Optional[str] = None,
-        offset: Optional[float] = None,
-        page_size: Optional[float] = None,
+        offset: Optional[int] = None,
+        page_size: Optional[int] = None,
         start_date: Optional[str] = None,
         namespace: Optional[str] = None,
         login_as: Optional[str] = None,
